Improved user input structure. Added input checking. Jointed the different libraries together
Added a new module called 'pointer_utils' that contains a function that veryfies a pointer by printing an error if it's equal to NULL. I added this pointer to every necessary function in this programm. I also improved the user input 'module' in main.c so the programm doesn't go rambo if you type in a letter. There was also a bug befor, that if you typed in 2 or 3 that the programm would say that you made a bad input. Fixed that too.

Signed-off-by: Ferdinand Leybach <ferdinand@leybach.de>